requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==3.0.1
PyJWT==2.8.0
stripe==8.7.0
python-dotenv==1.0.1
Flask-CORS==4.0.0
pytest==7.4.0
pytest-flask==1.2.0
pytest-mock==3.11.1
Flask-Mail==0.9.1